worked on admin theme and list and from
