+++ src/components/common/Footer.tsx
@@ -1,5 +1,22 @@
+import { Typography, Link } from "@mui/material";
+
 const Footer = (): JSX.Element => {
-    return <h1>test</h1>;
+  function Copyright(props: any) {
+    return (
+      <Typography variant="body2" color="text.secondary" align="center" {...props}>
+        {'Copyright © '}
+        <Link color="inherit" href="https://mui.com/">
+          Your Website
+        </Link>{' '}
+        {new Date().getFullYear()}
+        {'.'}
+      </Typography>
+    );
+  }
+
+  return <>
+    <Copyright sx={{ pt: 4 }} />
+  </>;
   };
   
 export default Footer;
