--- src/components/dashboard/Dashboard.tsx.preview
@@ -1,1 +1,0 @@
-<DashboardContent />
