--- src/components/dashboard/Title.tsx
@@ -1,14 +1,0 @@
-import * as React from 'react';
-import Typography from '@mui/material/Typography';
-
-interface TitleProps {
-  children?: React.ReactNode;
-}
-
-export default function Title(props: TitleProps) {
-  return (
-    <Typography component="h2" variant="h6" color="primary" gutterBottom>
-      {props.children}
-    </Typography>
-  );
-}
