--- src/components/home/Home.tsx
@@ -1,5 +1,0 @@
-const Home = (): JSX.Element => {
-  return <h1>test</h1>;
-};
-
-export default Home;
